move migration code into dal
