+++ server/persistence/relational/dal.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/jinzhu/gorm"
 	"github.com/offen/offen/server/persistence"
+	gormigrate "gopkg.in/gormigrate.v1"
 )
 
 func (r *relationalDatabase) createEvent(e *Event) error {
@@ -302,4 +303,39 @@ func (r *relationalDatabase) updateAccountUserRelationship(a *AccountUserRelatio
 		return fmt.Errorf("dal: error updating account user relationship: %w", err)
 	}
 	return nil
+}
+
+func (r *relationalDatabase) applyMigrations() error {
+	m := gormigrate.New(r.db, gormigrate.DefaultOptions, []*gormigrate.Migration{
+		{
+			ID: "001_add_one_time_keys",
+			Migrate: func(db *gorm.DB) error {
+				type AccountUserRelationship struct {
+					RelationshipID                    string `gorm:"primary_key"`
+					UserID                            string
+					AccountID                         string
+					PasswordEncryptedKeyEncryptionKey string
+					EmailEncryptedKeyEncryptionKey    string
+					// this is the field introducted in this migration
+					OneTimeEncryptedKeyEncryptionKey string
+				}
+				return db.AutoMigrate(&AccountUserRelationship{}).Error
+			},
+			Rollback: func(db *gorm.DB) error {
+				return db.Table("account_user_relationships").DropColumn("one_time_encrypted_key_encryption_key").Error
+			},
+		},
+	})
+
+	m.InitSchema(func(tx *gorm.DB) error {
+		return tx.AutoMigrate(
+			&Event{},
+			&Account{},
+			&User{},
+			&AccountUser{},
+			&AccountUserRelationship{},
+		).Error
+	})
+
+	return m.Migrate()
 }
